move all text gen to SourceFormatterGenerator
